Update server setup, configuration and database handling
This update refactors the server setup to wrap the websocket server around an HTTP server. It also adds new environment variables and replaces a WebSocketServer requirement with an HTTP requirement in the 'www' files. The database handling was improved with the possibility to connect to a MongoDB, and a new event model schema was created for the MongoDB setup. In-memory event store was kept for cases when MongoDB is not available.